+++ theories/heap_lang/proph_erasure.v
@@ -40,6 +40,7 @@ Fixpoint erase_expr (e : expr) : expr :=
   | Case e0 e1 e2 => Case (erase_expr e0) (erase_expr e1) (erase_expr e2)
   | Fork e => Fork (erase_expr e)
   | AllocN e1 e2 => AllocN (erase_expr e1) (erase_expr e2)
+  | Free e => Free (erase_expr e)
   | Load e => Load (erase_expr e)
   | Store e1 e2 => Store (erase_expr e1) (erase_expr e2)
   | CmpXchg e0 e1 e2 => CmpXchg (erase_expr e0) (erase_expr e1) (erase_expr e2)
@@ -93,6 +94,7 @@ Fixpoint erase_ectx_item (Ki : ectx_item) : list ectx_item :=
   | CaseCtx e1 e2 => [CaseCtx (erase_expr e1) (erase_expr e2)]
   | AllocNLCtx v2 => [AllocNLCtx (erase_val v2)]
   | AllocNRCtx e1 => [AllocNRCtx (erase_expr e1)]
+  | FreeCtx => [FreeCtx]
   | LoadCtx => [LoadCtx]
   | StoreLCtx v2 => [StoreLCtx (erase_val v2)]
   | StoreRCtx e1 => [StoreRCtx (erase_expr e1)]
@@ -115,7 +117,8 @@ Definition erase_ectx (K : ectx heap_ectx_lang) : ectx heap_ectx_lang :=
 
 Definition erase_tp (tp : list expr) : list expr := erase_expr <$> tp.
 
-Definition erase_heap (h : gmap loc val) : gmap loc val := erase_val <$> h.
+Definition erase_heap (h : gmap loc (option val)) : gmap loc (option val) :=
+  (λ ov : option val, erase_val <$> ov) <$> h.
 
 Definition erase_state (σ : state) : state :=
   {| heap := erase_heap (heap σ); used_proph_id := ∅ |}.
@@ -189,17 +192,22 @@ Qed.
 Lemma lookup_erase_heap_None h l : erase_heap h !! l = None ↔ h !! l = None.
 Proof. rewrite lookup_fmap; by destruct (h !! l). Qed.
 
-Lemma lookup_erase_heap h l : erase_heap h !! l = erase_val <$> h !! l.
+Lemma lookup_erase_heap h l : erase_heap h !! l = (λ ov, erase_val <$> ov) <$> h !! l.
 Proof. by rewrite lookup_fmap. Qed.
 
-Lemma erase_heap_insert h l v :
-  erase_heap (<[l := v]> h) = <[l := erase_val v]> (erase_heap h).
+Lemma erase_heap_insert_Some h l v :
+  erase_heap (<[l := Some v]> h) = <[l := Some $ erase_val v]> (erase_heap h).
+Proof.
+  by rewrite /erase_heap fmap_insert.
+Qed.
+Lemma erase_heap_insert_None h l v :
+  erase_heap (<[l := None]> h) = <[l := None]> (erase_heap h).
 Proof.
   by rewrite /erase_heap fmap_insert.
 Qed.
 
-Lemma fmap_heap_array f l vs :
-  f <$> heap_array l vs = heap_array l (f <$> vs).
+Lemma fmap_heap_array (f : val → val) l vs :
+  (λ ov : option val, f <$> ov) <$> heap_array l vs = heap_array l (f <$> vs).
 Proof.
   revert l; induction vs as [|v vs IHvs]; intros l;
     first by rewrite /= fmap_empty.
@@ -246,48 +254,57 @@ Proof.
     first econstructor; try setoid_rewrite <- lookup_erase_heap_None;
       rewrite ?erase_heap_insert /=; eauto using erase_state_init.
 Qed.
+Lemma erased_head_step_head_step_Free l v σ :
+  erase_heap (heap σ) !! l = Some (Some v) →
+  head_steps_to_erasure_of (Free #l) σ #()
+   {| heap := <[l:=None]> (erase_heap (heap σ)); used_proph_id := ∅ |} [].
+Proof.
+  intros Hl.
+  rewrite lookup_erase_heap in Hl.
+  destruct (heap σ !! l) as [[|]|] eqn:?; simplify_eq/=.
+  eexists _, _, _, _; simpl; split; first econstructor; repeat split; eauto.
+  rewrite /state_upd_heap /erase_state /= erase_heap_insert_None //.
+Qed.
 Lemma erased_head_step_head_step_Load l σ v :
-  erase_heap (heap σ) !! l = Some v →
+  erase_heap (heap σ) !! l = Some (Some v) →
   head_steps_to_erasure_of (! #l) σ v (erase_state σ) [].
 Proof.
   intros Hl.
   rewrite lookup_erase_heap in Hl.
-  destruct (heap σ !! l) eqn:?; simplify_eq/=.
+  destruct (heap σ !! l) as [[|]|] eqn:?; simplify_eq/=.
   eexists _, _, _, _; simpl; split; first econstructor; eauto.
 Qed.
-
-Lemma erased_head_step_head_step_Store l v σ :
-  is_Some (erase_heap (heap σ) !! l) →
-  head_steps_to_erasure_of (#l <- v) σ #()
-   {| heap := <[l:=erase_val v]> (erase_heap (heap σ)); used_proph_id := ∅ |} [].
+Lemma erased_head_step_head_step_Store l v w σ :
+  erase_heap (heap σ) !! l = Some (Some v) →
+  head_steps_to_erasure_of (#l <- w) σ #()
+   {| heap := <[l:=Some $ erase_val w]> (erase_heap (heap σ)); used_proph_id := ∅ |} [].
 Proof.
   intros Hl.
   rewrite lookup_erase_heap in Hl.
-  destruct (heap σ !! l) eqn:?; simplify_eq/=;
-           last by apply is_Some_None in Hl.
+  destruct (heap σ !! l) as [[|]|] eqn:?; simplify_eq/=.
   eexists _, _, _, _; simpl; split; first econstructor; repeat split; eauto.
-    by rewrite /state_upd_heap /erase_state /= erase_heap_insert.
+  rewrite /state_upd_heap /erase_state /= erase_heap_insert_Some //.
 Qed.
 Lemma erased_head_step_head_step_CmpXchg l v w σ vl :
-  erase_heap (heap σ) !! l = Some vl →
+  erase_heap (heap σ) !! l = Some (Some vl) →
   vals_compare_safe vl (erase_val v) →
   head_steps_to_erasure_of
     (CmpXchg #l v w) σ
     (vl, #(bool_decide (vl = erase_val v)))%V
     (if bool_decide (vl = erase_val v)
-     then {| heap := <[l:=erase_val w]> (erase_heap (heap σ));
+     then {| heap := <[l:=Some $ erase_val w]> (erase_heap (heap σ));
              used_proph_id := ∅ |}
      else erase_state σ) [].
 Proof.
   intros Hl Hvl.
   rewrite lookup_erase_heap in Hl.
-  destruct (heap σ !! l) as [u|] eqn:?; simplify_eq/=.
+  destruct (heap σ !! l) as [[u|]|] eqn:?; simplify_eq/=.
   rewrite -> vals_compare_safe_erase in Hvl.
   destruct (decide (u = v)) as [->|Hneq].
   - eexists _, _, _, _; simpl; split.
     { econstructor; eauto. }
     rewrite !bool_decide_eq_true_2; eauto using erase_val_inj_iff; [].
-    rewrite -?erase_heap_insert.
+    rewrite -?erase_heap_insert_Some.
     split_and!; auto.
   - eexists _, _, _, _; simpl; split.
     { econstructor; eauto. }
@@ -295,17 +312,17 @@ Proof.
     by rewrite erase_val_inj_iff.
 Qed.
 Lemma erased_head_step_head_step_FAA l n m σ :
-  erase_heap (heap σ) !! l = Some #n →
+  erase_heap (heap σ) !! l = Some (Some #n) →
   head_steps_to_erasure_of
     (FAA #l #m) σ #n
-    {| heap := <[l:= #(n + m)]> (erase_heap (heap σ));
+    {| heap := <[l:= Some #(n + m)]> (erase_heap (heap σ));
        used_proph_id := ∅ |} [].
 Proof.
   intros Hl.
   rewrite lookup_erase_heap in Hl.
-  destruct (heap σ !! l) as [[[]| | | |]|] eqn:?; simplify_eq/=.
+  destruct (heap σ !! l) as [[[[]| | | |]|]|] eqn:?; simplify_eq/=.
   repeat econstructor; first by eauto.
-  by rewrite /state_upd_heap /erase_state /= erase_heap_insert.
+  by rewrite /state_upd_heap /erase_state /= erase_heap_insert_Some.
 Qed.
 
 (** When the erased program makes a head step, so does the original program. *)
@@ -329,6 +346,7 @@ Proof.
   eauto using erased_head_step_head_step_rec,
     erased_head_step_head_step_NewProph,
     erased_head_step_head_step_AllocN,
+    erased_head_step_head_step_Free,
     erased_head_step_head_step_Load,
     erased_head_step_head_step_Store,
     erased_head_step_head_step_CmpXchg,
@@ -615,7 +633,7 @@ Proof.
 Qed.
 
 Lemma head_step_erased_prim_step_CmpXchg v1 v2 σ l vl:
-  heap σ !! l = Some vl →
+  heap σ !! l = Some (Some vl) →
   vals_compare_safe vl v1 →
   ∃ e2' σ2' ef', prim_step (CmpXchg #l (erase_val v1)
                              (erase_val v2)) (erase_state σ) [] e2' σ2' ef'.
@@ -667,8 +685,16 @@ Proof.
   intros; rewrite lookup_erase_heap_None; eauto.
 Qed.
 
+Lemma head_step_erased_prim_step_free σ l v :
+  heap σ !! l = Some (Some v) →
+  ∃ e2' σ2' ef', prim_step (Free #l) (erase_state σ) [] e2' σ2' ef'.
+Proof.
+  intros Hw. do 3 eexists; apply head_prim_step; econstructor.
+  rewrite /erase_state /state_upd_heap /= lookup_erase_heap Hw; eauto.
+Qed.
+
 Lemma head_step_erased_prim_step_load σ l v:
-  heap σ !! l = Some v →
+  heap σ !! l = Some (Some v) →
   ∃ e2' σ2' ef', prim_step (! #l) (erase_state σ) [] e2' σ2' ef'.
 Proof.
   do 3 eexists; apply head_prim_step; econstructor.
@@ -676,17 +702,16 @@ Proof.
   by destruct lookup; simplify_eq.
 Qed.
 
-Lemma head_step_erased_prim_step_store σ l v:
-  is_Some (heap σ !! l) →
-  ∃ e2' σ2' ef', prim_step (#l <- erase_val v) (erase_state σ) [] e2' σ2' ef'.
+Lemma head_step_erased_prim_step_store σ l v w :
+  heap σ !! l = Some (Some v) →
+  ∃ e2' σ2' ef', prim_step (#l <- erase_val w) (erase_state σ) [] e2' σ2' ef'.
 Proof.
-  intros [w Hw].
-  do 3 eexists; apply head_prim_step; econstructor.
+  intros Hw. do 3 eexists; apply head_prim_step; econstructor.
   rewrite /erase_state /state_upd_heap /= lookup_erase_heap Hw; eauto.
 Qed.
 
 Lemma head_step_erased_prim_step_FAA σ l n n':
-  heap σ !! l = Some #n →
+  heap σ !! l = Some (Some #n) →
   ∃ e2' σ2' ef', prim_step (FAA #l #n') (erase_state σ) [] e2' σ2' ef'.
 Proof.
   intros Hl.
@@ -709,6 +734,7 @@ Proof.
                 head_step_erased_prim_step_un_op,
                 head_step_erased_prim_step_bin_op,
                 head_step_erased_prim_step_allocN,
+                head_step_erased_prim_step_free,
                 head_step_erased_prim_step_load,
                 head_step_erased_prim_step_store,
                 head_step_erased_prim_step_FAA;